requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-floppyforms==1.7.0
Django==1.10.3
django-localflavor-us==1.1
django-localflavor==1.3
psycopg2==2.6.2
selenium==3.0.1
whitenoise==3.2.2
